requirements.txt
streamlit==1.33.0
transformers==4.33.3
google-generativeai==0.5.0
ffmpeg-python==0.2.0
torch==2.2.2
